refactor(web): separate the jwt token logic into a separate repository and mediator
