+++ web/src/mediators/index.ts
@@ -1,1 +1,2 @@
 export * from './auth.mediator';
+export * from './token.mediator';
