+++ web/src/mediators/token.mediator.ts
@@ -1,0 +1,21 @@
+import { inject, Injectable } from '@angular/core';
+import { TokenRepository } from '@repositories/auth';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class TokenMediator {
+  readonly #tokenRepository = inject(TokenRepository);
+
+  getTokens() {
+    return this.#tokenRepository.getTokens();
+  }
+
+  getAccessToken() {
+    return this.#tokenRepository.getAccessToken();
+  }
+
+  getRefreshToken() {
+    return this.#tokenRepository.getRefreshToken();
+  }
+}
