+++ web/src/repositories/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './dto';
 export * from './auth.repository';
+export * from './token.repository';
