+++ web/src/repositories/auth/token.repository.ts
@@ -1,0 +1,36 @@
+import { Injectable } from '@angular/core';
+import { TokensModel } from '@models/tokens.model';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class TokenRepository {
+  storeTokens(tokens: TokensModel): void {
+    if (tokens.accessToken) {
+      localStorage.setItem('accessToken', tokens.accessToken);
+    }
+    if (tokens.refreshToken) {
+      localStorage.setItem('refreshToken', tokens.refreshToken);
+    }
+  }
+
+  getTokens(): TokensModel {
+    const accessToken = localStorage.getItem('accessToken');
+    const refreshToken = localStorage.getItem('refreshToken');
+
+    return new TokensModel(accessToken, refreshToken);
+  }
+
+  getAccessToken(): string | null {
+    return localStorage.getItem('accessToken');
+  }
+
+  getRefreshToken(): string | null {
+    return localStorage.getItem('refreshToken');
+  }
+
+  clearTokens(): void {
+    localStorage.removeItem('accessToken');
+    localStorage.removeItem('refreshToken');
+  }
+}
